refactor(test): extract shared R test helpers into common module
Move duplicated `with_r`, `ensure_r_initialized`, and
`ld_library_path_is_set` helpers from startup.rs, visibility.rs, and
help.rs into tests/common/mod.rs to avoid behavior drift and make
future R integration tests easier to add.

Also fix the ellipses assertion comment in help.rs to accurately
describe what is being checked (the \dots macro regression only,
not braces).
