+++ crates/arf-harp/tests/startup.rs
@@ -1,53 +1,9 @@
 //! Integration tests for R startup hook functions.
 
+mod common;
+
 use arf_harp::{call_dot_first, call_dot_first_sys, eval_string};
-use once_cell::sync::OnceCell;
-use std::sync::Mutex;
-
-static R_LOCK: OnceCell<Mutex<()>> = OnceCell::new();
-
-fn ensure_r_initialized() -> bool {
-    static R_INITIALIZED: OnceCell<bool> = OnceCell::new();
-
-    *R_INITIALIZED.get_or_init(|| unsafe {
-        match arf_libr::initialize_r() {
-            Ok(()) => true,
-            Err(e) => {
-                eprintln!("Failed to initialize R: {}", e);
-                false
-            }
-        }
-    })
-}
-
-fn with_r<F, T>(f: F) -> Option<T>
-where
-    F: FnOnce() -> T,
-{
-    if !ensure_r_initialized() {
-        return None;
-    }
-    let lock = R_LOCK.get_or_init(|| Mutex::new(()));
-    // Use into_inner() to recover from a poisoned lock caused by a previous panic.
-    let _guard = lock.lock().unwrap_or_else(|e| e.into_inner());
-    Some(f())
-}
-
-/// Check that LD_LIBRARY_PATH includes the R library directory.
-/// Tests that require package loading (e.g. utils, methods) must be skipped
-/// when this is not set, because `initialize_r()` cannot re-exec the process
-/// as the binary does via `ensure_ld_library_path()`.
-fn ld_library_path_is_set() -> bool {
-    let Ok(lib_path) = arf_libr::find_r_library() else {
-        return false;
-    };
-    let Some(lib_dir) = lib_path.parent() else {
-        return false;
-    };
-    let lib_dir_str = lib_dir.to_string_lossy();
-    let current = std::env::var("LD_LIBRARY_PATH").unwrap_or_default();
-    current.split(':').any(|p| p == lib_dir_str.as_ref())
-}
+use common::{ld_library_path_is_set, with_r};
 
 #[test]
 fn test_call_dot_first_noop_when_undefined() {
